cleanup code use size_t type for memory amount

+++ bld/dip/codeview/c/cvmisc.c
@@ -48,7 +48,7 @@ unsigned DIPIMPENTRY( HandleSize )( handle_kind hk )
     return( Sizes[hk] );
 }
 
-dip_status DIPIMPENTRY( MoreMem )( unsigned size )
+dip_status DIPIMPENTRY( MoreMem )( size_t size )
 {
     /* unused parameters */ (void)size;
